feat(auth): implement role-based access control and user authentication feat(dashboard): create DashboardClient component for live monitoring feat(control-panel): enhance control panel with role-based actions
