+++ src/web/dashboard/types/next-auth.d.ts
@@ -1,0 +1,19 @@
+import type { DefaultSession } from 'next-auth';
+
+declare module 'next-auth' {
+  interface Session {
+    user: {
+      role?: string;
+    } & DefaultSession['user'];
+  }
+
+  interface User {
+    role?: string;
+  }
+}
+
+declare module 'next-auth/jwt' {
+  interface JWT {
+    role?: string;
+  }
+}
